fixed tests, moved custom matchers to util package
