+++ sdk/src/main/AndroidManifest.xml
@@ -87,7 +87,8 @@
             android:screenOrientation="portrait"
             android:theme="@style/Theme.InfoMercadoPagoTheme.NoActionbar" />
         <activity
-            android:name=".ResultActivity" />
+            android:name=".ResultActivity"
+            android:theme="@style/Theme.MercadoPagoTheme"/>
     </application>
 
 </manifest>
